fixed typo, changed flow of game to go from intro->characterSelection->MainGameplay

+++ src/support/debugger/support/shapes/CircleShapeDefine.java
@@ -1,34 +1,35 @@
 package support.debugger.support.shapes;
 
+import support.Vec2d;
 import support.debugger.collisions.CircleShape;
 import support.debugger.support.Display;
 import support.debugger.support.Vec2f;
 
 public class CircleShapeDefine extends CircleShape {
 	
-	public CircleShapeDefine(Vec2f center, float radius) {
+	public CircleShapeDefine(Vec2d center, float radius) {
 		super(center, radius);
 	}
 
 	@Override
-	public void move(Vec2f distance) {
+	public void move(Vec2d distance) {
 		center = center.plus(distance);
 		bindToCanvas();
 	}
 	
 	@Override
 	public void bindToCanvas() {
-		Vec2f distance = new Vec2f(0);
+		Vec2d distance = new Vec2d(0);
 		if(center.x - radius < 0) {
-			distance = distance.plus(new Vec2f(radius - center.x, 0));
+			distance = distance.plus(new Vec2d(radius - center.x, 0));
 		} else if(center.x + radius >= Display.getStageWidth()) {
-			distance = distance.plus(new Vec2f(Display.getStageWidth() - center.x - radius, 0));
+			distance = distance.plus(new Vec2d(Display.getStageWidth() - center.x - radius, 0));
 		}
 		
 		if(center.y - radius < 0) {
-			distance = distance.plus(new Vec2f(0, radius - center.y));
+			distance = distance.plus(new Vec2d(0, radius - center.y));
 		} else if(center.y + radius >= Display.getStageHeight()) {
-			distance = distance.plus(new Vec2f(0, Display.getStageHeight() - center.y - radius));
+			distance = distance.plus(new Vec2d(0, Display.getStageHeight() - center.y - radius));
 		}
 		
 		center = center.plus(distance);
